Partially apply clang-tidy fixes we don't enforce yet (#8376)
* Partially apply clang-tidy fixes we don't use yet

- Put a bunch of stuff into anonymous namespaces
- Delete some redundant casts (e.g. casting an int to int)
- Add some const refs to avoid copies
- Remove meaningless inline qualifiers on in-class definitions and
constexpr functions
- Remove return-with-value from functions returning void
- Delete a little dead code
- Use std::min/max where appropriate
- Don't use a variable after std::forwarding it. It may have been moved
from.
- Use std::string::empty instead of comparing length to zero

* Undo unintentional formatting change

* Restore some necessary casts

* Add NOLINT to silence older clang-tidy

+++ src/FlattenNestedRamps.cpp
@@ -91,7 +91,7 @@ class FlattenRamps : public IRMutator {
                         c /= stride;
                     }
                     // Compute the number of elements loaded
-                    extent = (int)((max_constant_offset / stride) + 1);
+                    extent = (max_constant_offset / stride) + 1;
                 }
 
                 // If we're gathering from a very large range, it
